corrected prev/next pages path and fixed pep8 mistakes

+++ src/schemas/movies.py
@@ -14,8 +14,8 @@ class MovieBase(BaseModel):
     orig_title: str
     status: str
     orig_lang: str
-    budget: float
-    revenue: float
+    budget: int
+    revenue: int
     country: str
